# backend-app/configs/swagger_config.json
{
    "swagger": "2.0",
    "info": {
        "title": "Software Voting Machine API",
        "description": "This API allows us to manage all related endpoints to our software voting machine app.",
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/roles": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a new role",
                "description": "Creates a new role with the given data.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "role",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RoleCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Role created successfully",
                        "schema": {
                            "$ref": "#/definitions/RoleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "description": "Returns a list of all roles in the database.",
                "responses": {
                    "200": {
                        "description": "A list of roles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RoleResponse"
                            }
                        }
                    }
                }
            }
        },
        "/roles/{role_id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get a role by ID",
                "description": "Retrieves the role with the specified ID.",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role found",
                        "schema": {
                            "$ref": "#/definitions/RoleResponse"
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update a role by ID",
                "description": "Updates the role with the given ID using the provided data.",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "role",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RoleCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role updated successfully",
                        "schema": {
                            "$ref": "#/definitions/RoleResponse"
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Delete a role by ID",
                "description": "Deletes the role with the specified ID.",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role deleted successfully"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RoleCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the role"
                },
                "organisation_id": {
                    "type": "integer",
                    "description": "The ID of the organisation associated with the role"
                }
            },
            "required": [
                "name",
                "id"
            ]
        },
        "RoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "The unique identifier of the role"
                },
                "name": {
                    "type": "string",
                    "description": "The name of the role"
                },
                "organisation_id": {
                    "type": "integer",
                    "description": "The ID of the organisation associated with the role"
                }
            }
        },
...
    }
}